% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_plotting.R
\name{iterate_plotting}
\alias{iterate_plotting}
\title{Generate Multiple Plots with Fixed or Dynamic Axes
This function iterates over a dataset and generates multiple plots using a specified plotting function. It supports fixed or dynamic axis limits based on the overall dataset or additional arguments provided. Each subset of data is passed to the plotting function along with optional parameters.}
\usage{
iterate_plotting(
  df,
  fx,
  input = "standard",
  fixed_axes = TRUE,
  save = FALSE,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame containing the dataset with \code{x}, \code{y}, \code{x_label}, and \code{y_label} columns.}

\item{fx}{A function to generate individual plots. It should accept arguments \code{df_split}, \code{input}, \code{save},
and optionally \code{min_x}, \code{max_x}, \code{min_y}, \code{max_y}.}

\item{input}{A character string specifying the input type, either "standard" or "fraction". Default is "standard".}

\item{fixed_axes}{Logical. If TRUE, uses global axis limits for all plots. If FALSE, each plot can have dynamic axis limits. Default is TRUE.}

\item{save}{Logical. If TRUE, saves each plot generated by the \code{fx} function. Default is FALSE.}

\item{n_cores}{Integer. Specifies the number of cores to use for parallel processing. Default is NULL (automatically detects available cores).}

\item{...}{Additional arguments passed to \code{fx}, allowing custom axis limits (\code{min_x}, \code{max_x}, \code{min_y}, \code{max_y}).}
}
\value{
A list of plots generated by the \code{fx} function, each created based on a subset of \code{df}.
}
\description{
Generate Multiple Plots with Fixed or Dynamic Axes
This function iterates over a dataset and generates multiple plots using a specified plotting function. It supports fixed or dynamic axis limits based on the overall dataset or additional arguments provided. Each subset of data is passed to the plotting function along with optional parameters.
}
\examples{
# Example usage:
# plots <- iterate_plotting(df, fx = plot_phase_portrait, input = "standard", fixed_axes = TRUE)
}
