% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra.R
\name{lotka_volterra}
\alias{lotka_volterra}
\title{Lotka-Volterra Model Simulation
This function generates an example dataset for this package, simulating several trajectories using predator-prey equations.}
\usage{
lotka_volterra(
  alpha = 0.5,
  beta = 0.8,
  gamma = 1,
  delta = 0.8,
  min_x = 1,
  max_x = 10,
  bin_width_x = 1,
  min_y = 1,
  max_y = 5,
  bin_width_y = 1,
  min_time = 0,
  max_time = 25,
  time_interval = 0.1
)
}
\arguments{
\item{alpha}{Numeric. The growth rate of the prey population. Default is 1.1.}

\item{beta}{Numeric. The rate of predation, describing how frequently predators consume prey. Default is 0.4.}

\item{gamma}{Numeric. The mortality rate of the predator population. Default is 0.1.}

\item{delta}{Numeric. The reproduction rate of predators per prey consumed. Default is 0.01.}

\item{min_x}{Numeric. Minimum initial population of prey. Default is 10.}

\item{max_x}{Numeric. Maximum initial population of prey. Default is 50.}

\item{bin_width_x}{Numeric. Interval width for the prey population bins. Default is 10.}

\item{min_y}{Numeric. Minimum initial population of predators. Default is 5.}

\item{max_y}{Numeric. Maximum initial population of predators. Default is 25.}

\item{bin_width_y}{Numeric. Interval width for the predator population bins. Default is 5.}

\item{min_time}{Numeric. Starting time for the simulation. Default is 0.}

\item{max_time}{Numeric. Ending time for the simulation. Default is 50.}

\item{time_interval}{Numeric. Time increment for each simulation step. Default is 0.1.}
}
\value{
A data frame containing the time series of prey and predator populations for each initial condition
and parameter set. The output includes columns for time, prey and predator populations, and initial conditions.
}
\description{
This function simulates the Lotka-Volterra predator-prey model over a specified time range
and initial conditions for the populations of prey and predator species. It generates a table of
population dynamics for various starting conditions and parameter values.
}
\details{
The Lotka-Volterra model describes the dynamics of biological systems in which two species interact,
one as a predator and the other as prey. This function allows for exploring various initial conditions
and parameter values to study the system's behavior under different scenarios.
}
\examples{
# Run the Lotka-Volterra simulation with default parameters
results <- lotka_volterra()

# Run with custom parameters
results <- lotka_volterra(alpha = 1.2, beta = 0.5, gamma = 0.15, delta = 0.02)

}
