% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_portrait.R
\name{phase_portrait}
\alias{phase_portrait}
\title{Generate Phase Portrait Summary (Standard or Fraction-Based)}
\usage{
phase_portrait(
  df,
  bin_width = 0.25,
  min_bin_n = 1,
  input = "standard",
  save = FALSE
)
}
\arguments{
\item{df}{A data frame containing the following columns:
\itemize{
\item \code{condition}: Factor or character column indicating the experimental or observational condition.
\item \code{x}: Numeric column representing values for the x-axis variable.
\item \code{y}: Numeric column representing values for the y-axis variable.
\item \code{dx}: Numeric column representing the rate of change of \code{x}.
\item \code{dy}: Numeric column representing the rate of change of \code{y}.
\item \code{x_variable}, \code{y_variable}: Character or factor columns identifying the variables.
}}

\item{bin_width}{Integer specifying the number of bins to discretize \code{x} and \code{y} variables. Default is 4.}

\item{min_bin_n}{Integer specifying the minimum number of observations required in each bin for it to be included in the output. Default is 1.}

\item{input}{Character string specifying the input of data, either "standard" for standard binning or "fraction" for fraction-based binning.}

\item{save}{Logical, if TRUE, saves the resulting data frame as a compressed CSV file. Default is FALSE.}
}
\value{
A data frame with summarized phase portrait data containing:
\itemize{
\item \code{condition}, \code{x_variable}, \code{y_variable}: Grouping columns from the input.
\item \code{x}, \code{y}: Discretized values of \code{x} and \code{y} or fractional representation of \code{x} and total size as \code{y}.
\item \code{dx}, \code{dy}: Median values of \code{dx} and \code{dy} within each bin.
\item \code{n}: Number of observations in each bin.
\item \code{x_label}, \code{y_label}: Labels for x and y variables, adapted based on the input.
\item \code{dx_label}, \code{dy_label}: Labels for rate of change for x and y variables.
\item \code{input}: The input of the data ("standard" or "fraction").
}
}
\description{
This function generates a phase portrait by binning \code{x} and \code{y} variables into discrete intervals and calculating the median rate of change (\code{dx} and \code{dy}) within each bin. It supports both standard and fraction-based portraits, where \code{x} values can be shown as fractions of the total \code{(x + y)}, to visualize dynamics across binned states.
}
\examples{
# Example usage
phase_portrait(df, bin_width = 5, min_bin_n = 2, input = "standard")
phase_portrait(df, bin_width = 5, min_bin_n = 2, input = "fraction")
}
