% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_line.R
\name{plot_phase_line}
\alias{plot_phase_line}
\title{Plot Phase Lines in Phase Space
This function generates a plot of phase lines in phase space, showing changes in \code{x} and \code{y} values (change rate) along with the grouping variable. It supports both standard and fraction-based input types and provides options to adjust axis limits, customize the plot's color scale, and save the plot as a PDF.}
\arguments{
\item{df}{A data frame containing columns \code{x}, \code{y}, \code{n}, \code{group_var}, \code{condition}, \code{facet}, \code{x_variable}, and \code{y_variable}.}

\item{input}{Character string specifying the input type, either "standard" or "fraction". Default is "standard".}

\item{min_x, }{max_x Optional numeric values to set limits for the x-axis.}

\item{min_y, }{max_y Optional numeric values to set limits for the y-axis.}

\item{save}{Logical, if TRUE, saves the plot as a PDF file. Default is FALSE.}
}
\value{
A ggplot object representing phase lines in phase space.
}
\description{
Plot Phase Lines in Phase Space
This function generates a plot of phase lines in phase space, showing changes in \code{x} and \code{y} values (change rate) along with the grouping variable. It supports both standard and fraction-based input types and provides options to adjust axis limits, customize the plot's color scale, and save the plot as a PDF.
}
\examples{
plot_phase_line(df, input = "standard", min_x = -10, max_x = 10, min_y = -5, max_y = 5, save = TRUE)
}
