% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_portrait.R
\name{plot_phase_portrait}
\alias{plot_phase_portrait}
\title{Plot Phase Portrait}
\usage{
plot_phase_portrait(
  df,
  input = "standard",
  bin_width = 0.25,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL,
  save = FALSE,
  fix_coord = FALSE
)
}
\arguments{
\item{df}{A data frame containing \code{x}, \code{y}, \code{dx}, \code{dy}, \code{condition}, \code{x_variable}, and \code{y_variable}.}

\item{input}{A character string indicating the type of input. Defaults to \code{"standard"}.}

\item{bin_width}{Numeric. Adjusts the length of arrows. Default is 0.25.}

\item{min_x}{Numeric. Minimum x-axis limit for the plot. Default is NULL.}

\item{max_x}{Numeric. Maximum x-axis limit for the plot. Default is NULL.}

\item{min_y}{Numeric. Minimum y-axis limit for the plot. Default is NULL.}

\item{max_y}{Numeric. Maximum y-axis limit for the plot. Default is NULL.}

\item{save}{Logical, if TRUE saves the plot to a PDF file. Default is FALSE.}

\item{fix_coord}{Logical, if TRUE enforces a fixed aspect ratio. Default is FALSE.}
}
\value{
A ggplot object showing the phase portrait.
}
\description{
This function generates a phase portrait plot, visualizing the dynamics of two variables (\code{x} and \code{y}) with arrows representing the rate of change (\code{dx} and \code{dy}). The plot uses color to indicate the speed of change and can optionally enforce axis limits, aspect ratios, and save the plot to a file.
}
\examples{
plot_phase_portrait(df, input = "standard", min_x = -10, max_x = 10, min_y = -5, max_y = 5, save = TRUE)
}
