% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stream.R
\name{plot_stream}
\alias{plot_stream}
\title{Plot Phase Streamline
This function creates a phase streamline plot, visualizing the dynamics of two variables (\code{x} and \code{y}) across conditions. It supports both standard and fraction-based input types, with options for binning, custom axis limits, fixed aspect ratio, and saving the plot.}
\usage{
plot_stream(
  df,
  input = "standard",
  bin_width = 0.25,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL,
  save = FALSE,
  fix_coord = FALSE,
  res = 25
)
}
\arguments{
\item{df}{A data frame containing columns \code{x}, \code{y}, \code{dx}, \code{dy}, \code{condition}, \code{x_variable}, and \code{y_variable}.}

\item{input}{A character string indicating the type of input, either \code{"standard"} or \code{"fraction"}. Default is \code{"standard"}.}

\item{bin_width}{Integer specifying the number of bins for discretizing \code{x} and \code{y}. Default is 0.25.}

\item{min_x}{Numeric. Minimum x-axis limit for the plot. Default is NULL.}

\item{max_x}{Numeric. Maximum x-axis limit for the plot. Default is NULL.}

\item{min_y}{Numeric. Minimum y-axis limit for the plot. Default is NULL.}

\item{max_y}{Numeric. Maximum y-axis limit for the plot. Default is NULL.}

\item{save}{Logical, if TRUE, saves the plot as a PDF file. Default is FALSE.}
}
\value{
A ggplot object showing the phase streamline.
}
\description{
Plot Phase Streamline
This function creates a phase streamline plot, visualizing the dynamics of two variables (\code{x} and \code{y}) across conditions. It supports both standard and fraction-based input types, with options for binning, custom axis limits, fixed aspect ratio, and saving the plot.
}
\examples{
plot_stream(df, input = "standard", bin_width = 4, min_x = 0, max_x = 1, min_y = 0, max_y = 1, save = TRUE)
}
