% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot Trajectory in Phase Space}
\usage{
plot_trajectory(
  df,
  input = "standard",
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL,
  save = FALSE,
  fix_coord = FALSE
)
}
\arguments{
\item{df}{A dataframe containing columns for coordinates (\code{x}, \code{y}), pseudotime, and directional endpoints (\code{x1}, \code{x3}, \code{y1}, \code{y3}). Optionally includes \code{input} (plot label), \code{condition} (facet label), \code{x_variable}, and \code{y_variable}.}

\item{input}{Character. Input type for the plot label. Default is "standard". If \code{df$input} exists, it will override this parameter.}

\item{min_x}{Numeric. Minimum x-axis limit for the plot. Default is NULL.}

\item{max_x}{Numeric. Maximum x-axis limit for the plot. Default is NULL.}

\item{min_y}{Numeric. Minimum y-axis limit for the plot. Default is NULL.}

\item{max_y}{Numeric. Maximum y-axis limit for the plot. Default is NULL.}

\item{save}{Logical. If TRUE, saves the plot as a PDF file. Default is FALSE.}
}
\value{
A ggplot object of the trajectory plot, with customizable color and axis limits.
}
\description{
This function creates a phase space trajectory plot, visualizing \code{x} and \code{y} values along pseudotime. It connects \code{x1} to \code{x} and \code{x} to \code{x3} as well as \code{y1} to \code{y} and \code{y} to \code{y3}, representing directional flows with arrows. This enables a clear depiction of how each point in phase space progresses over time.
}
\details{
This function creates a phase space trajectory plot where points and paths are visualized with arrows to indicate directionality. It uses pseudotime as a color gradient to show progression through phase space. Faceting by the \code{condition} column allows for separate plots by specified condition.
}
\examples{
# Example usage:
df <- data.frame(
  x = c(1, 2, 3), y = c(3, 2, 1),
  x1 = c(1.1, 2.1, 3.1), x3 = c(0.9, 1.9, 2.9),
  y1 = c(3.1, 2.1, 1.1), y3 = c(2.9, 1.9, 0.9),
  pseudotime = c(0.1, 0.5, 0.9),
  condition = c("A", "B", "A"),
  x_variable = "X Axis", y_variable = "Y Axis",
  input = "custom_label"
)
plot_trajectory(df, min_x = 0, max_x = 5, min_y = 0, max_y = 5)
}
