% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_trajectory.R
\name{pseudotime_trajectory}
\alias{pseudotime_trajectory}
\title{Calculate Pseudotime Trajectory Summary Statistics}
\usage{
pseudotime_trajectory(df, input = "standard", save = FALSE)
}
\arguments{
\item{df}{A data frame containing the following columns:
\itemize{
\item \code{condition}: Factor or character column indicating the experimental or observational condition.
\item \code{pseudotime}: Numeric column indicating pseudotime or a sequential progression.
\item \code{x}: Numeric column representing values for the x-axis variable.
\item \code{y}: Numeric column representing values for the y-axis variable.
\item \code{x_variable}: Character or factor column identifying the name of the \code{x} variable.
\item \code{y_variable}: Character or factor column identifying the name of the \code{y} variable.
}}

\item{input}{Either "standard" or "fraction" to specify the type of analysis.}

\item{save}{Logical, if TRUE saves the resulting data frame as a CSV file. Default is FALSE.}
}
\value{
A data frame containing the calculated summary statistics for each combination
of \code{condition} and \code{pseudotime}. Columns in the returned data frame include:
\itemize{
\item \code{x1}: 25th percentile of \code{x} values.
\item \code{y1}: 25th percentile of \code{y} values.
\item \code{x3}: 75th percentile of \code{x} values.
\item \code{y3}: 75th percentile of \code{y} values.
\item \code{x}: Median of \code{x} values.
\item \code{y}: Median of \code{y} values.
\item \code{n}: Number of observations in each group.
}
}
\description{
This function calculates summary statistics (median and quantiles) for \code{x} and \code{y} variables along a pseudotime trajectory within each condition, computing the 25th and 75th percentiles, as well as the median for each group, to analyze trends in \code{x} and \code{y} over pseudotime.
}
\examples{
# Example usage:
pseudotime_trajectory(df)
}
